deep-presentation-client/pyproject.toml
[tool.poetry]
name = "deeppresentation-client"
version = "0.1.0"
description = ""
authors = ["Jakub Melzacki <melzacki.jakub@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
